chore: minor refactoring
