+++ src/service/ping.rs
