+++ src/service/register_circuit.rs
@@ -2,7 +2,7 @@ use keccak_hash::keccak;
 
 use crate::types::{register_circuit::RegisterCircuitRequest, register_circuit::RegisterCircuitResponse};
 
-pub fn register_circuit(data: RegisterCircuitRequest) -> RegisterCircuitResponse {
+pub fn register_circuit_exec(data: RegisterCircuitRequest) -> RegisterCircuitResponse {
     // 1. Dump whatever you need in db/queue 
     // 2. Just return the keccak hash of the data.vkey at this point
     // Keccak hash borshified vkey for now
